+++ testdata/bindings/bindings.go
@@ -30,6 +30,7 @@ func Exports(r bind.Registry) {
 	r.Func(sample.MakeEmbedder)
 	r.Func(sample.TakesTime)
 	r.Func(sample.MakeStamped)
+	r.Func(sample.MaybeStamp)
 	r.Func(sample.Advance)
 	r.Func(sample.Replay)